Refactor filtering ability to util function and add exit button
