+++ frontend/src/util/Utility.ts
@@ -7,6 +7,7 @@ import { setRoom } from '../redux/Room';
 import { setCurrentUser } from '../redux/User';
 import { setGame } from '../redux/Game';
 import { AppDispatch } from '../redux/Store';
+import { Problem, SelectableProblem } from '../api/Problem';
 
 // Require validator for identifiers as no types are provided.
 const validateIdentifier = require('valid-identifier');
@@ -90,3 +91,19 @@ export const getAuthHttpHeader = (token: string) => ({
     Authorization: token,
   },
 });
+
+export const problemMatchesFilterText = (problem: Problem | SelectableProblem, filterText: string): boolean => {
+  const texts = filterText.toLowerCase().split(',');
+
+  // Filter by name, difficulty, and tags (multiple queries separated by commas)
+  for (let i = 0; i < texts.length; i += 1) {
+    const text = texts[i].trim();
+    if (!problem.name.toLowerCase().includes(text)
+      && !problem.difficulty.toLowerCase().includes(text)
+      && !problem.problemTags.some((tag) => tag.name.toLowerCase().includes(text))) {
+      return false;
+    }
+  }
+
+  return true;
+};
